pyproject.toml
[project]
name = "nightshift"
version = "0.1.0"
description = "Autonomous overnight research agent for TOTO time series fine-tuning"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.0",
    "paramiko>=3.4.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tabulate>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gpu = [
    "toto-ts",
    "torch>=2.5.0",
    "lightning>=2.3.0",
    "datasets>=2.17.0",
    "gluonts[torch]>=0.16.0",
]
